+++ src/Forms/FormField.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+
+using System.Collections.Generic;
 
 namespace Forms
 {
@@ -10,21 +11,28 @@ namespace Forms
     ///     Inspired by ION spec (see http://ionwg.org/draft-ion.html#form-fields for more details)
     /// </para>
     /// </summary>
-    /// <remarks>
-    /// </remarks>
     [JsonObject]
+#if NET5_0_OR_GREATER
+    public record FormField
+#else
     public class FormField
+#endif
     {
         private IEnumerable<FormFieldOption> _options;
 
         /// <summary>
         /// indicates whether or not the field value may be modified or submitted to a linked resource location. 
         /// </summary>
-        /// <remarks>
-        /// 
-        /// </remarks>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public bool? Enabled { get; set; }
+        public bool? Enabled
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Type of the field
@@ -36,68 +44,156 @@ namespace Forms
         /// Description of the field
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public string Label { get; set; }
+        public string Label
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Name of the field that should be submitted
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public string Name { get; set; }
+        public string Name
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Regular expression that the field should be validated against.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public string Pattern { get; set; }
+        public string Pattern
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Short hint that described the expected value of the field.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public string Placeholder { get; set; }
+        public string Placeholder
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates if the field must be submitted
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public bool? Required { get; set; }
+        public bool? Required
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates the maximum length of the value
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public int? MaxLength { get; set; }
+        public int? MaxLength
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates the minimum length of the value
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public int? MinLength { get; set; }
+        public int? MinLength
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates that value must be greater than or equal to the specified <see cref="Min"/> value
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public float? Min { get; set; }
+        public float? Min
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates that value must be less than or equal to the specified <see cref="Max"/> value
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public float? Max { get; set; }
+        public float? Max
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates whether or not the field value is considered sensitive information 
         /// and should be kept secret.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public bool? Secret { get; set; }
+        public bool? Secret
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// a string description of the field that may be used to enhance usability.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public string Description { get; set; }
+        public string Description
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// List of options
@@ -105,7 +201,11 @@ namespace Forms
         public IEnumerable<FormFieldOption> Options
         {
             get => _options;
+#if NET5_0_OR_GREATER
+            init
+#else
             set
+#endif
             {
                 _options = value;
                 Type = FormFieldType.Array;