+++ src/Forms/Link.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Forms
 {
@@ -9,20 +10,38 @@ namespace Forms
     /// <remarks>
     ///     Inspired by ION spec (see http://ionwg.org/draft-ion.html#links for more details)
     /// </remarks>
+#if NETSTANDARD2_1_OR_GREATER
+    public record Link
+#else
     public class Link
+#endif
     {
         /// <summary>
         /// Url of the resource the current <see cref="Link"/> points to.
         /// </summary>
-        public string Href { get; set; }
+        public string Href
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Relations of the resource that <see cref="Link"/> points to with the current resource
         /// </summary>
 #if NETSTANDARD
-        public IEnumerable<string> Relations { get; set; } 
+        public IEnumerable<string> Relations
+        {
+            get => _relations.Distinct();
+            set => _relations = value?.Distinct() ?? Enumerable.Empty<string>();
+        }
+
+        private IEnumerable<string> _relations;
 #else
-        public IReadOnlySet<string> Relations { get; set; }
+        public IReadOnlySet<string> Relations { get; init; }
 #endif
 
         /// <summary>
@@ -30,7 +49,15 @@ namespace Forms
         /// </summary>
         /// <remarks>
         /// </remarks>
-        public string Method { get; set; }
+        public string Method
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Title associated with the link
@@ -38,7 +65,15 @@ namespace Forms
         /// <remarks>
         /// Should be a friendly name suitable to used in a HTML a tag.
         /// </remarks>
-        public string Title { get; set; }
+        public string Title
+        {
+            get;
+#if NET5_0_OR_GREATER
+            init;
+#else
+            set;
+#endif
+        }
 
         /// <summary>
         /// Indicates if the current <see cref="Href"/> is a template url
@@ -56,7 +91,11 @@ namespace Forms
         /// </summary>
         public Link()
         {
+#if NET5_0_OR_GREATER
             Relations = new HashSet<string>();
+#else
+            _relations = new HashSet<string>();
+#endif
         }
 
         /// <inheritdoc/>